Allow statically linking modules
The goal was to enable this without adding any speacial cases into the main
module loading code. The way we do this is that for a statically linked module,
we generate a shared module stub that basically just searches for the symbols in
the main executable and defers to them. This means that statically linked
module's create and destroy functions must be prefixed by something unique.